chore: delete next-intl
