+++ package.json
@@ -25,7 +25,6 @@
     "lucide-react": "^0.219.0",
     "next": "13.3.0",
     "next-contentlayer": "^0.3.1",
-    "next-intl": "^2.13.0-beta.2",
     "next-sitemap": "^4.0.7",
     "next-themes": "^0.2.1",
     "react": "18.2.0",
