+++ src/app/blog/ViewCounter.tsx
