+++ src/app/blog/page.tsx
@@ -8,8 +8,8 @@ export const metadata = {
 function Blog() {
   return (
     <div style={{ opacity: 1, transform: 'none' }}>
-      <h1>Weekly</h1>
-      <Article category="weekly" />
+      <h1>文章</h1>
+      <Article category="blog" />
     </div>
   );
 }
