+++ src/app/page.tsx
@@ -1,0 +1,18 @@
+export default function Home() {
+  return (
+    <>
+      <h1>👋&nbsp;大家好</h1>
+      <p>
+        我是 <b>YoungLe</b>，
+        一名后知后觉热爱编程和游戏的前端开发工程师。同时也热衷于捣鼓各种新鲜的工具和技术。
+      </p>
+      <p>
+        我也会写一些 开源项目，此外还喜欢在 博客 上记录我的开发工作和日常生活。
+      </p>
+      <p>
+        我曾尝试使用多个博客应用，但写下有价值的内容并不多。这次我打算使用
+        Next.js 实现自己的博客，并计划花一些时间好好沉淀和书写。
+      </p>
+    </>
+  );
+}
