+++ src/app/weekly/[slug]/page.tsx
