+++ src/components/mdx.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import * as React from 'react';
-import { Link } from 'next-intl';
+import Link from 'next/link';
 import Image from 'next/image';
 import { useMDXComponent } from 'next-contentlayer/hooks';
