--- src/i18n/i18n.server.ts
@@ -1,5 +1,0 @@
-export async function getMessages({ locale }: RootParams) {
-  const { default: messages } = await import(`@/messages/${locale}.json`);
-
-  return messages as IntlMessages;
-}
