--- src/messages/en.json
@@ -1,39 +1,0 @@
-{
-  "home": {
-    "welcome": "Hi! There",
-    "paragraph1": "I am YoungLe，a front-end development engineer who loves programming and games as an afterthought",
-    "paragraph2": "I also write about <github-link>open source</github-link> projects, and in addition I like to <blog-link>blog</blog-link> about my development work and daily life.",
-    "paragraph3": "I've tried using several blogging applications, but not much valuable content has been written down. This time I'm going to implement my own blog using Next.js and plan to spend some time sinking my teeth into it and writing about it.",
-    "tools": "Tools",
-    "software": "Software",
-    "software-li1": "This website is built with <Next.js-link>Next.js</Next.js-link>, and <tailwindcss-link>Tailwindcss</tailwindcss-link>. I use <vercel-link>Vercel</vercel-link> for hosting",
-    "software-li2": "This blog post is written in <mdx-link>MDX</mdx-link> and parsed using <contentLayer-link>ContentLayer</contentLayer-link>.",
-    "software-li3": "<b>Coding</b>: <vsCode-link>VS Code</vsCode-link> with <github-theme-link>GitHub Theme</github-theme-link>",
-    "software-li4": "<b>Notes</b>: <notion-link>Notion</notion-link>",
-    "software-li5": "<b>Browser</b>: <arc-link>ARC</arc-link>",
-    "software-li6": "<b>Music</b>: <spotify-link>Spotify</spotify-link>",
-    "software-li7": "<b>Project Manage</b>: <linear-link>Linear</linear-link>",
-    "hardware": "Hardware",
-    "hardware-li1": "<b>Coding PC</b>: MacBook Pro 13 M1 (2021)",
-    "hardware-li2": "<b>Gaming</b>: Switch, PS5",
-    "openSource": "Open Source",
-    "chatgpt": "Use as the official ChatGPT, with even more powerful features.",
-    "hackintosh": "i5-8400+ASRock Z370M-ITX/AC Hackintosh Build"
-  },
-  "navbar": {
-    "home": "home",
-    "about": "about Me",
-    "blog": "blog",
-    "contact": "contact",
-    "dark": "Dark",
-    "light": "Light",
-    "weekly": "weekly"
-  },
-  "blog": {
-    "title": "Blog",
-    "views": "views"
-  },
-  "footer": {
-    "totalViews": "Total Views"
-  }
-}
