--- src/messages/zh.json
@@ -1,39 +1,0 @@
-{
-  "home": {
-    "welcome": "大家好",
-    "paragraph1": "我是 <b>YoungLe</b>， 一名后知后觉热爱编程和游戏的前端开发工程师。同时也热衷于捣鼓各种新鲜的工具和技术。",
-    "paragraph2": "我也会写一些 <github-link>开源项目</github-link>，此外还喜欢在 <blog-link>博客</blog-link> 上记录我的开发工作和日常生活。",
-    "paragraph3": "我曾尝试使用多个博客应用，但写下有价值的内容并不多。这次我打算使用 Next.js 实现自己的博客，并计划花一些时间好好沉淀和书写。",
-    "tools": "工具",
-    "software": "软件",
-    "software-li1": "这个项目使用 <Next.js-link>Next.js</Next.js-link> 和 <tailwindcss-link>Tailwindcss</tailwindcss-link> 构建，部署在 <vercel-link>Vercel</vercel-link> 上",
-    "software-li2": "博客文章使用 <mdx-link>MDX</mdx-link> 编写，使用 <contentLayer-link>ContentLayer</contentLayer-link> 解析",
-    "software-li3": "<b>Coding</b>: <vsCode-link>VS Code</vsCode-link> 和 <github-theme-link>GitHub Theme</github-theme-link>",
-    "software-li4": "<b>Notes</b>: <notion-link>Notion</notion-link>",
-    "software-li5": "<b>Browser</b>: <arc-link>ARC</arc-link>",
-    "software-li6": "<b>Music</b>: <spotify-link>Spotify</spotify-link>",
-    "software-li7": "<b>Project Manage</b>: <linear-link>Linear</linear-link>",
-    "hardware": "硬件",
-    "hardware-li1": "<b>Coding PC</b>: MacBook Pro 13 M1 (2021)",
-    "hardware-li2": "<b>Gaming</b>: Switch, PS5",
-    "openSource": "开源项目",
-    "chatgpt": "像官方 ChatGPT 一样使用，再加入更强大的功能",
-    "hackintosh": "i5-8400 + ASRock Z370M-ITX/AC 黑苹果构建  "
-  },
-  "navbar": {
-    "home": "主页",
-    "about": "关于Me",
-    "blog": "博客",
-    "contact": "联系我",
-    "dark": "深色",
-    "light": "浅色",
-    "weekly": "周刊"
-  },
-  "blog": {
-    "title": "博客",
-    "views": "阅读"
-  },
-  "footer": {
-    "totalViews": "总访问量"
-  }
-}
